Add files via upload
This project is based on Bootstrap and React basics. In this project user enters some text then the text can be converted into Uppercase or lowercase. Also, we can count the number of words in the text.
To execute this project download the node modules and other essentials because this rar file only contains Public and Src Folders.